agent/requirements.txt
requests>=2.31,<3
psycopg2-binary>=2.9,<3
numpy>=1.24,<2
openai>=1.30,<2
fastapi>=0.111,<1
uvicorn[standard]>=0.29,<1
pydantic>=2.7,<3
